bot/requirements.txt
ac-library-python @ git+https://github.com/not522/ac-library-python@58f324ec020d57191e7b9e4957b0c5feb5ed3aff
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
appdirs==1.4.4
asyncio-dgram==2.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.5.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
discord.py==2.4.0
dnspython==2.7.0
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-compute==1.25.0
google-cloud-core==2.4.1
google-cloud-monitoring==2.27.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
Mako==1.3.2
MarkupSafe==2.1.5
mcstatus==11.1.1
multidict==6.1.0
online-judge-api-client==10.10.1
online-judge-template-generator==4.8.1
online-judge-tools==11.5.1
outcome==1.3.0.post0
platformdirs==4.2.0
ply==3.11
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.17.1
rsa==4.9
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
toml==0.10.2
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.2.0
wsproto==1.2.0
yapf==0.40.2
yarl==1.18.3
zipp==3.17.0
